Fix: incorrect results when merging deltas of concurrent writes
Reduce the number of iterations of a property based test

Add new property based test for the RGA

Test the merge of deltas of concurrent insertions.

Fix the RGA merge

There was a problem when calculating the position to insert.

Reduce the number of iterations of a property based test